embassy_executor
Module raw

Raw executor.

This module exposes “raw” Executor and Task structs for more low level control.

WARNING: here be dragons!

Using this module requires respecting subtle safety contracts. If you can, prefer using the safe executor wrappers and the embassy_executor::task macro, which are fully safe.

Structs

AvailableTask

An uninitialized TaskStorage.

Executor

Raw executor.

TaskPool

Raw storage that can hold up to N tasks of the same type.

TaskRef

This is essentially a &'static TaskStorage<F> where the type of the future has been erased.

TaskStorage

Raw storage in which a task can be spawned.

Functions

task_from_waker

Get a task pointer from a waker.

wake_task

Wake a task by TaskRef.

wake_task_no_pend

Wake a task by TaskRef without calling pend.